+++ src/Factory/User/One.php
@@ -1,10 +1,8 @@
 <?php
 
-namespace Recca0120\Socialite\One;
+namespace Recca0120\Socialite\Factory\User;
 
-use Recca0120\Socialite\AbstractUser;
-
-class User extends AbstractUser
+class One extends AbstractUser
 {
     /**
      * The user's access token.
